Crate fedimint_server

Modules

atomic_broadcast

This module implements fedimints custom atomic broadcast abstraction. A such, it is responsible for ordering serialized items in the form of byte vectors. The Broadcast is able to recover from a crash at any time via a backup that it maintains in the servers fedimint_core::db::Database. In Addition, it stores the history of accepted items in the form of signed session outcomes in the database as well in order to catch up fellow guardians which have been offline for a prolonged period of time.

config

Fedimint toplevel config

consensus

The actual implementation of consensus

envs

metrics

multiplexed

Implementation of multiplexed peer connections

net

Networking for mint-to-mint and client-to-mint communiccation

Functions

get_config

run

run_config_gen