Module fedimint_server::api

Structs

ApiVersionSet

Set of api versions for each component (core + modules)

ClientConfigDownloadToken

Allows a client to download the config

DynGlobalApi

DynModuleApi

FederationError

An API request error when calling an entire federation

FederationStatus

The status of a server, including how it views its peers

InviteCode

Information required for client to construct WsFederationApi instance

PeerResponse

PeerStatus

StatusResponse

WsFederationApi

Mint API client that will try to run queries against all peers expecting equal results from at least min_eq_results of them. Peers that return differing results are returned as a peer faults list.

Enums

OutputOutcomeError

PeerConnectionStatus

PeerError

An API request error when calling a single federation peer

ServerStatus

The state of the server returned via APIs

Traits

FederationApiExt

An extension trait allowing to making federation-wide API call on top IFederationApi.

GlobalFederationApi

The API for the global (non-module) endpoints

IFederationApi

An API (module or global) that can query a federation

IGlobalFederationApi

Trait marker for the global (non-module) endpoints

IModuleFederationApi

Trait marker for the module (non-global) endpoints

JsonRpcClient

Type Definitions

FederationResult

JsonRpcResult

PeerResult