Struct fedimint_server::api::WsFederationApi

pub struct WsFederationApi<C = Client> {
    peer_ids: BTreeSet<PeerId, Global>,
    peers: Arc<Vec<FederationPeer<C>, Global>, Global>,
    module_id: Option<u16>,
}

Mint API client that will try to run queries against all peers expecting equal results from at least min_eq_results of them. Peers that return differing results are returned as a peer faults list.

Fields
peer_ids: BTreeSet<PeerId, Global>peers: Arc<Vec<FederationPeer<C>, Global>, Global>module_id: Option<u16>
Implementations

impl WsFederationApi<Client>

pub fn new(peers: Vec<(PeerId, SafeUrl), Global>) -> WsFederationApi<Client>

Creates a new API client

pub fn from_config(config: &ClientConfig) -> WsFederationApi<Client>

Creates a new API client from a client config

pub fn from_invite_code(info: &[InviteCode]) -> WsFederationApi<Client>

Creates a new API client from a invite code, assumes they are in peer id order

impl<C> WsFederationApi<C>

pub fn peers(&self) -> Vec<PeerId, Global>

pub fn new_with_client( peers: Vec<(PeerId, SafeUrl), Global> ) -> WsFederationApi<C>

Creates a new API client