Trait fedimint_server::api::IFederationApi

pub trait IFederationApi: Debug + MaybeSend + MaybeSync {
    fn all_members(&self) -> &BTreeSet<PeerId, Global>;
    fn request_raw<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        peer_id: PeerId,
        method: &'life1 str,
        params: &'life2 [Value]
    ) -> Pin<Box<dyn Future<Output = Result<Value, Error>> + Send + 'async_trait, Global>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait;
}

Required Methods

fn all_members(&self) -> &BTreeSet<PeerId, Global>

List of all federation members for the purpose of iterating each member in the federation.

The underlying implementation is resonsible for knowing how many and PeerIds of each. The caller of this interface most probably have some idea as well, but passing this set across every API call to the federation would be inconvenient.

fn request_raw<'life0, 'life1, 'life2, 'async_trait>(
&'life0 self,
peer_id: PeerId,
method: &'life1 str,
params: &'life2 [Value]
) -> Pin<Box<dyn Future<Output = Result<Value, Error>> + Send + 'async_trait, Global>>where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

Make request to a specific federation member by peer_id