Type Alias fedimint_server::api::SerdeOutputOutcome

pub type SerdeOutputOutcome = SerdeModuleEncoding<DynOutputOutcome>;

Aliased Type

struct SerdeOutputOutcome(Vec<u8, Global>, PhantomData<DynOutputOutcome>);

Fields
0: Vec<u8, Global>1: PhantomData<DynOutputOutcome>
Implementations

impl<T> SerdeModuleEncoding<T>where T: Encodable + Decodable + 'static,

pub fn try_into_inner( &self, modules: &ModuleRegistry<Decoder, DecodingMode> ) -> Result<T, DecodeError>

pub fn try_into_inner_known_module_kind( &self, decoder: &Decoder ) -> Result<T, DecodeError>

In cases where we know exactly which module kind we expect but don’t have access to all decoders this function can be used instead.

Note that it just assumes the decoded module instance id to be valid since it cannot validate against the decoder registry. The lack of access to a decoder registry also makes decoding structs impossible that themselves contain module dyn-types (e.g. a module output containing a fedimint transaction).