fedimint_server

Module config

Fedimint toplevel config

Modules

dkg

dkg_g1

dkg_g2

io

peer_handle

setup

Structs

ClientConfig

Total client config

ConfigGenParams

All the parameters necessary for generating the ServerConfig during setup

ConfigGenSettings

All the info we configure prior to config gen starting

FederationId

The federation id is a copy of the authentication threshold public key of the federation

GlobalClientConfig

Federation-wide client config

JsonWithKind

serde_json::Value that must contain kind: String field

ModuleInitRegistry

PeerIrohEndpoints

PeerUrl

ServerConfig

All the serializable configuration for the fedimint server

ServerConfigConsensus

ServerConfigLocal

ServerConfigPrivate

ServerModuleConfig

Config for the server-side of a particular Federation module

ServerModuleConsensusConfig

Enums

P2PMessage

Constants

DEFAULT_BROADCAST_ROUNDS_PER_SESSION 🔒

DEFAULT_BROADCAST_ROUND_DELAY_MS 🔒

Consensus broadcast settings that result in 3 minutes session time

DEFAULT_MAX_CLIENT_CONNECTIONS

The default maximum open connections the API can handle

DEFAULT_TEST_BROADCAST_ROUNDS_PER_SESSION 🔒

DEFAULT_TEST_BROADCAST_ROUND_DELAY_MS 🔒

Consensus broadcast settings that result in 10 seconds session time

Traits

TypedServerModuleConfig

Module (server side) config, typed

Functions

default_broadcast_rounds_per_session 🔒

legacy_consensus_config_hash