Type Alias fedimint_server::config::ServerModuleConfigGenParamsRegistry

pub type ServerModuleConfigGenParamsRegistry = ModuleRegistry<ConfigGenModuleParams, ()>;

Registry that contains the config gen params for all modules

Aliased Type

struct ServerModuleConfigGenParamsRegistry {
    inner: BTreeMap<u16, (ModuleKind, ConfigGenModuleParams), Global>,
    state: (),
}

Fields
inner: BTreeMap<u16, (ModuleKind, ConfigGenModuleParams), Global>state: ()
Implementations

impl ModuleRegistry<ConfigGenModuleParams, ()>

pub fn attach_config_gen_params<T>( &mut self, id: u16, kind: ModuleKind, gen: T ) -> &mut ModuleRegistry<ConfigGenModuleParams, ()>where T: ModuleInitParams,

impl<M, State> ModuleRegistry<M, State>

pub fn new( iter: impl IntoIterator<Item = (u16, ModuleKind, M)> ) -> ModuleRegistry<M, State>where State: Default,

Create Self from an iterator of pairs

pub fn iter_modules(&self) -> impl Iterator<Item = (u16, &ModuleKind, &M)>

Return an iterator over all module data

pub fn get(&self, id: u16) -> Option<&M>

Get module data by instance id

pub fn get_with_kind(&self, id: u16) -> Option<&(ModuleKind, M)>

Get module data by instance id, including ModuleKind

impl<M, State> ModuleRegistry<M, State>where M: Debug,

pub fn get_expect(&self, id: u16) -> &M

Return the module data belonging to the module identified by the supplied module_id

Panics

If the module isn’t in the registry

pub fn register_module(&mut self, id: u16, kind: ModuleKind, module: M)

Add a module to the registry