Module fedimint_server::consensus::aleph_bft

Modules

backup

data_provider

finalization_handler

keychain

network

spawner

Structs

Message

The majority of these messages need to be delivered to the intended Recipient in order for aleph bft to make progress. However, alpeh bft does not assume a reliable network layer and implements all necessary retry logic. Therefore, the network layer can discard a message if its intended recipient is offline.

Enums

Recipient

This enum defines the intended recipient of a Message.

Functions

to_node_index

to_peer_id