Struct fedimint_server::consensus::api::ConsensusApi

pub struct ConsensusApi {
    pub cfg: ServerConfig,
    pub db: Database,
    pub modules: ServerModuleRegistry,
    pub client_cfg: ClientConfig,
    pub submission_sender: Sender<ConsensusItem>,
    pub shutdown_sender: Sender<Option<u64>>,
    pub connection_status_channels: Arc<RwLock<BTreeMap<PeerId, PeerConnectionStatus>>>,
    pub last_ci_by_peer: Arc<RwLock<BTreeMap<PeerId, u64>>>,
    pub supported_api_versions: SupportedApiVersionsSummary,
}

Fields
cfg: ServerConfig

Our server configuration

db: Database

Database for serving the API

modules: ServerModuleRegistry

Modules registered with the federation

client_cfg: ClientConfig

Cached client config

submission_sender: Sender<ConsensusItem>

For sending API events to consensus such as transactions

shutdown_sender: Sender<Option<u64>>connection_status_channels: Arc<RwLock<BTreeMap<PeerId, PeerConnectionStatus>>>last_ci_by_peer: Arc<RwLock<BTreeMap<PeerId, u64>>>supported_api_versions: SupportedApiVersionsSummary
Implementations

impl ConsensusApi

pub fn api_versions_summary(&self) -> &SupportedApiVersionsSummary

pub async fn submit_transaction( &self, transaction: Transaction ) -> Result<TransactionId, TransactionError>

pub async fn await_transaction( &self, txid: TransactionId ) -> (Vec<ModuleInstanceId>, DatabaseTransaction<'_, Committable>)

pub async fn await_output_outcome( &self, outpoint: OutPoint ) -> Result<SerdeModuleEncoding<DynOutputOutcome>>

pub async fn session_count(&self) -> u64

pub async fn await_signed_session_outcome( &self, index: u64 ) -> SignedSessionOutcome

pub async fn session_status(&self, session_index: u64) -> SessionStatus

pub async fn get_federation_status(&self) -> ApiResult<FederationStatus>

fn shutdown(&self, index: Option<u64>)

async fn get_federation_audit(&self) -> ApiResult<AuditSummary>

async fn get_guardian_config_backup( &self, password: String ) -> ApiResult<GuardianConfigBackup>

Uses the in-memory config to write a config backup tar archive that guardians can download. Private keys are encrypted with the guardian password, so it should be safe to store anywhere, this also means the backup is useless without the password.

async fn handle_backup_request<'s, 'dbtx, 'a>( &'s self, dbtx: &'dbtx mut DatabaseTransaction<'a>, request: SignedBackupRequest ) -> Result<(), ApiError>

async fn handle_recover_request( &self, dbtx: &mut DatabaseTransaction<'_>, id: PublicKey ) -> Option<ClientBackupSnapshot>