Struct fedimint_server::consensus::engine::ConsensusEngine

pub struct ConsensusEngine {

    pub modules: ServerModuleRegistry,
    pub db: Database,
    pub federation_api: DynGlobalApi,
    pub cfg: ServerConfig,
    pub submission_receiver: Receiver<ConsensusItem>,
    pub shutdown_receiver: Receiver<Option<u64>>,
    pub last_ci_by_peer: Arc<RwLock<BTreeMap<PeerId, u64>>>,
    pub self_id_str: String,
    pub peer_id_str: Vec<String>,
    pub connection_status_channels: Arc<RwLock<BTreeMap<PeerId, PeerConnectionStatus>>>,
    pub task_group: TaskGroup,
    pub data_dir: PathBuf,
    pub checkpoint_retention: u64,
    pub p2p_bind_addr: SocketAddr,

}

Runs the main server consensus loop

Fields
modules: ServerModuleRegistrydb: Databasefederation_api: DynGlobalApicfg: ServerConfigsubmission_receiver: Receiver<ConsensusItem>shutdown_receiver: Receiver<Option<u64>>last_ci_by_peer: Arc<RwLock<BTreeMap<PeerId, u64>>>self_id_str: String

Just a string version of cfg.local.identity for performance

peer_id_str: Vec<String>

Just a string version of peer ids for performance

connection_status_channels: Arc<RwLock<BTreeMap<PeerId, PeerConnectionStatus>>>task_group: TaskGroupdata_dir: PathBufcheckpoint_retention: u64p2p_bind_addr: SocketAddr
Implementations

impl ConsensusEngine

fn num_peers(&self) -> NumPeers

fn identity(&self) -> PeerId

pub async fn run(self) -> Result<()>

pub async fn run_single_guardian(&self, task_handle: TaskHandle) -> Result<()>

pub async fn run_consensus( &self, p2p_bind_addr: SocketAddr, task_handle: TaskHandle, ) -> Result<()>

async fn confirm_server_config_consensus_hash(&self) -> Result<()>

pub async fn run_session( &self, connections: ReconnectPeerConnections<Message>, session_index: u64, ) -> Result<()>

pub async fn complete_signed_session_outcome( &self, session_index: u64, ordered_unit_receiver: Receiver<OrderedUnit>, signature_sender: Sender<Option<SchnorrSignature>>, timestamp_receiver: Receiver<(Instant, u64)>, ) -> Result<SignedSessionOutcome>

fn decoders(&self) -> ModuleDecoderRegistry

pub async fn pending_accepted_items(&self) -> Vec<AcceptedItem>

pub async fn complete_session( &self, session_index: u64, signed_session_outcome: SignedSessionOutcome, )

fn db_checkpoints_dir(&self) -> PathBuf

Returns the full path where the database checkpoints are stored.

fn initialize_checkpoint_directory(&self, current_session: u64) -> Result<()>

Creates the directory within the data directory for storing the database checkpoints or deletes checkpoints before current_session - checkpoint_retention.

fn checkpoint_database(&self, session_index: u64)

Creates a backup of the database in the checkpoint directory. These checkpoints can be used to restore the database in case the federation falls out of consensus (recommended for experts only).

fn delete_old_database_checkpoint( &self, session_index: u64, checkpoint_dir: &Path, ) -> Result<()>

Deletes the database checkpoint directory equal to session_index - checkpoint_retention

pub async fn process_consensus_item( &self, session_index: u64, item_index: u64, item: ConsensusItem, peer: PeerId, ) -> Result<()>

async fn process_consensus_item_with_db_transaction( &self, dbtx: &mut DatabaseTransaction<'_>, consensus_item: ConsensusItem, peer_id: PeerId, ) -> Result<()>

async fn request_signed_session_outcome( &self, federation_api: &DynGlobalApi, index: u64, ) -> SignedSessionOutcome

async fn get_finished_session_count(&self) -> u64

Returns the number of sessions already saved in the database. This count does not include the currently running session.