Struct fedimint_server::consensus::server::ConsensusServer

pub struct ConsensusServer {

    pub task_group: TaskGroup,
    pub consensus: FedimintConsensus,
    pub api_receiver: Peekable<ReceiverStream<ApiEvent>>,
    pub connections: PeerConnections<EpochMessage>,
    pub cfg: ServerConfig,
    pub hbbft: HoneyBadger<Vec<SerdeConsensusItem>, PeerId>,
    pub api: DynGlobalApi,
    pub other_peers: BTreeSet<PeerId>,
    pub rejoin_at_epoch: HashMap<u64, HashSet<PeerId>>,
    pub latest_contribution_by_peer: Arc<RwLock<HashMap<PeerId, u64>>>,
    pub pending_forced_epochs: u64,
    pub last_processed_epoch: Option<SignedEpochOutcome>,
    pub decoders: ModuleDecoderRegistry,

}

Runs the main server consensus loop

Fields
task_group: TaskGroup

TaskGroup that is running the server

consensus: FedimintConsensus

Delegate for processing consensus information

api_receiver: Peekable<ReceiverStream<ApiEvent>>

Receives event notifications from the API (triggers epochs)

connections: PeerConnections<EpochMessage>

P2P connections for running consensus

cfg: ServerConfig

Our configuration

hbbft: HoneyBadger<Vec<SerdeConsensusItem>, PeerId>

Runs the HBBFT consensus algorithm

api: DynGlobalApi

Used to make API calls to our peers

other_peers: BTreeSet<PeerId>

The list of all other peers

rejoin_at_epoch: HashMap<u64, HashSet<PeerId>>

If Some then we restarted and look for the epoch to rejoin at

latest_contribution_by_peer: Arc<RwLock<HashMap<PeerId, u64>>>

Under the HBBFT consensus algorithm, this will track the latest epoch message received by each peer and when it was received

pending_forced_epochs: u64

Number of pending forced epochs (requested by peers to help join consensus faster)

last_processed_epoch: Option<SignedEpochOutcome>

Tracks the last epoch outcome from consensus

decoders: ModuleDecoderRegistry

Used for decoding module specific-values

Implementations

impl ConsensusServer

pub async fn new( cfg: ServerConfig, db: Database, module_inits: ServerModuleInitRegistry, task_group: &mut TaskGroup ) -> Result<Self>

Creates a server with real network and no delays

pub async fn new_with( cfg: ServerConfig, db: Database, module_inits: ServerModuleInitRegistry, connector: PeerConnector<EpochMessage>, delay_calculator: DelayCalculator, task_group: &mut TaskGroup ) -> Result<Self>

Creates a server that can simulate network and delays

Initializes modules and runs any database migrations

pub async fn run_consensus(self, task_handle: TaskHandle) -> Result<()>

Loop run_consensus_epoch until shut down

pub async fn start_consensus(&mut self)

Starts consensus by skipping to the last saved epoch history and triggering a new epoch

fn next_epoch_to_process(&self) -> u64

Returns the next epoch that we need to process, based on our saved history

pub async fn process_outcome( &mut self, last_outcome: HbbftConsensusOutcome ) -> Result<(), EpochVerifyError>

Requests, verifies and processes history from peers

last_outcome - The consensus outcome (unprocessed), we’re trying to process.

pub async fn run_consensus_epoch( &mut self, override_proposal: Option<ConsensusProposal>, rng: &mut (impl RngCore + CryptoRng + Clone + 'static) ) -> Result<Vec<HbbftConsensusOutcome>>

The main consensus function:

Await a new proposal event or receiving a proposal from peers

Send the ConsensusProposal to peers

Run HBBFT until a ConsensusOutcome can be returned

async fn process_events_then_propose( &mut self, override_proposal: Option<ConsensusProposal> ) -> ConsensusProposal

async fn force_process_epoch(&mut self, outcome: EpochOutcome)

async fn handle_step( &mut self, step: Step<Vec<SerdeConsensusItem>, PeerId> ) -> Cancellable<Vec<HbbftConsensusOutcome>>

Handles one step of the HBBFT algorithm, sending messages to peers and parsing any outcomes contained in the step

async fn propose_epoch( &mut self, proposal: ConsensusProposal, rng: &mut (impl RngCore + CryptoRng + Clone + 'static) ) -> Cancellable<Step<Vec<SerdeConsensusItem>, PeerId>>

async fn await_next_epoch(&mut self) -> Result<EpochTriggerEvent>

fn start_next_epoch(&self, msg: &(PeerId, EpochMessage)) -> bool

async fn handle_message( &mut self, msg: (PeerId, EpochMessage) ) -> Cancellable<Vec<HbbftConsensusOutcome>>

Runs a single HBBFT consensus step

async fn rejoin_at_epoch(&mut self, epoch: u64, peer: PeerId)

If we are rejoining and received a threshold of messages from the same epoch, then skip to that epoch. Give ourselves a buffer of NUM_EPOCHS_REJOIN_AHEAD so we can ensure we receive enough HBBFT messages to produce an outcome.

async fn request_rejoin(&mut self, epochs_to_run: u64)

Sends a rejoin request to all peers, indicating the number of epochs we want them to create