Module fedimint_server::core

Fedimint Core API (common) module interface

Fedimint supports externally implemented modules.

This (Rust) module defines common interoperability types and functionality that is used on both client and sever side.

Modules

client

server

Fedimint Core Server module interface

Macros

module_dyn_newtype_impl_encode_decode

Implement Encodable and Decodable for a “module dyn newtype”

module_plugin_trait_define

Define a “plugin” trait

newtype_impl_eq_passthrough_with_instance_id

plugin_types_trait_impl_common

Implements the necessary traits for all associated types of a FederationServer module.

Structs

Decoder

Decoder for module associated types

DecoderBuilder

DynInput

An owned, immutable input to a Transaction

DynModuleConsensusItem

An owned, immutable output of a Transaction before it was finalized

DynOutput

An owned, immutable output of a Transaction

DynOutputOutcome

An owned, immutable output of a Transaction before it was finalized

KeyPair

Opaque data structure that holds a keypair consisting of a secret and a public key.

ModuleKind

A type of a module

Enums

FinalizationError

Constants

LEGACY_HARDCODED_INSTANCE_ID_LN

LEGACY_HARDCODED_INSTANCE_ID_MINT

LEGACY_HARDCODED_INSTANCE_ID_WALLET

MODULE_INSTANCE_ID_GLOBAL

Special ID we use for global dkg

Traits

IInput

Something that can be an DynInput in a Transaction

IModuleConsensusItem

IOutput

Something that can be an DynOutput in a Transaction

IOutputOutcome

Input

IntoDynInstance

A type that has a Dyn*, type erased version of itself

ModuleConsensusItem

Output

OutputOutcome

Type Definitions

ModuleInstanceId

Module instance ID