Trait fedimint_server::core::Decoder

pub trait Decoder: Debug + Send + Sync + 'static {
    type Input: Input;
    type Output: Output;
    type OutputOutcome: OutputOutcome;
    type ConsensusItem: ModuleConsensusItem;

    // Required methods
    fn decode_input(&self, r: &mut dyn Read) -> Result<Self::Input, DecodeError>;
    fn decode_output(
        &self,
        r: &mut dyn Read
    ) -> Result<Self::Output, DecodeError>;
    fn decode_output_outcome(
        &self,
        r: &mut dyn Read
    ) -> Result<Self::OutputOutcome, DecodeError>;
    fn decode_consensus_item(
        &self,
        r: &mut dyn Read
    ) -> Result<Self::ConsensusItem, DecodeError>;
}

Module Decoder trait

Static-polymorphism version of IDecoder

All methods are static, as the decoding code is supposed to be instance-independent, at least until we start to support modules with overriden ModuleInstanceIds

Required Associated Types

type Input: Input

type Output: Output

type OutputOutcome: OutputOutcome

type ConsensusItem: ModuleConsensusItem

Required Methods

fn decode_input(&self, r: &mut dyn Read) -> Result<Self::Input, DecodeError>

Decode Input compatible with this module, after the module key prefix was already decoded

fn decode_output(&self, r: &mut dyn Read) -> Result<Self::Output, DecodeError>

Decode Output compatible with this module, after the module key prefix was already decoded

fn decode_output_outcome( &self, r: &mut dyn Read ) -> Result<Self::OutputOutcome, DecodeError>

Decode OutputOutcome compatible with this module, after the module key prefix was already decoded

fn decode_consensus_item( &self, r: &mut dyn Read ) -> Result<Self::ConsensusItem, DecodeError>

Decode ConsensusItem compatible with this module, after the module key prefix was already decoded