Trait fedimint_server::core::IDecoder

pub trait IDecoder: Debug {
    // Required methods
    fn decode_input(
        &self,
        r: &mut dyn Read,
        instance_id: u16
    ) -> Result<DynInput, DecodeError>;
    fn decode_output(
        &self,
        r: &mut dyn Read,
        instance_id: u16
    ) -> Result<DynOutput, DecodeError>;
    fn decode_output_outcome(
        &self,
        r: &mut dyn Read,
        instance_id: u16
    ) -> Result<DynOutputOutcome, DecodeError>;
    fn decode_consensus_item(
        &self,
        r: &mut dyn Read,
        instance_id: u16
    ) -> Result<DynModuleConsensusItem, DecodeError>;
}

Required Methods

fn decode_input( &self, r: &mut dyn Read, instance_id: u16 ) -> Result<DynInput, DecodeError>

Decode Input compatible with this module, after the module key prefix was already decoded

fn decode_output( &self, r: &mut dyn Read, instance_id: u16 ) -> Result<DynOutput, DecodeError>

Decode Output compatible with this module, after the module key prefix was already decoded

fn decode_output_outcome( &self, r: &mut dyn Read, instance_id: u16 ) -> Result<DynOutputOutcome, DecodeError>

Decode OutputOutcome compatible with this module, after the module key prefix was already decoded

fn decode_consensus_item( &self, r: &mut dyn Read, instance_id: u16 ) -> Result<DynModuleConsensusItem, DecodeError>

Decode ConsensusItem compatible with this module, after the module key prefix was already decoded