Type Alias fedimint_server::core::ModuleInstanceId

pub type ModuleInstanceId = u16;

Module instance ID

This value uniquely identifies a single instance of a module in a federation.

In case a single ModuleKind is instantiated twice (rare, but possible), each instance will have a different id.

Note: We have used this type differently before, assuming each u16 uniquly identifies a type of module in question. This function will move to a ModuleKind type which only identifies type of a module (mint vs wallet vs ln, etc)