Module fedimint_server::macros

Macros

_dyn_newtype_define_inner

_dyn_newtype_define_with_instance_id_inner

_dyn_newtype_impl_deref_mut

dyn_newtype_define

Define “dyn newtype” (a newtype over dyn Trait)

dyn_newtype_display_passthrough

Implements the Display trait for dyn newtypes whose traits implement Display

dyn_newtype_impl_dyn_clone_passhthrough

Implement Clone on a “dyn newtype”

erased_eq_no_instance_id

module_plugin_dyn_newtype_clone_passhthrough

module_plugin_dyn_newtype_define

Define a “module plugin dyn-newtype” which is like a standard “dyn newtype”, but with associated “module_instance_id”.

module_plugin_dyn_newtype_display_passthrough

module_plugin_dyn_newtype_encode_decode

Implement Encodable and Decodable for a “module dyn newtype”

module_plugin_dyn_newtype_eq_passthrough

module_plugin_static_trait_define

Define a “plugin” trait

plugin_types_trait_impl_common

Implements the necessary traits for all associated types of a FederationServer module.

plugin_types_trait_impl_config

Implements the necessary traits for all configuration related types of a FederationServer module.