Type Alias fedimint_server::module::registry::ModuleDecoderRegistry

pub type ModuleDecoderRegistry = ModuleRegistry<Decoder, DecodingMode>;

Collection of decoders belonging to modules, typically obtained from a ModuleRegistry

Aliased Type

struct ModuleDecoderRegistry {
    inner: BTreeMap<u16, (ModuleKind, Decoder), Global>,
    state: DecodingMode,
}

Fields
inner: BTreeMap<u16, (ModuleKind, Decoder), Global>state: DecodingMode
Implementations

impl<M, State> ModuleRegistry<M, State>

pub fn new( iter: impl IntoIterator<Item = (u16, ModuleKind, M)> ) -> ModuleRegistry<M, State>where State: Default,

Create Self from an iterator of pairs

pub fn iter_modules(&self) -> impl Iterator<Item = (u16, &ModuleKind, &M)>

Return an iterator over all module data

pub fn get(&self, id: u16) -> Option<&M>

Get module data by instance id

pub fn get_with_kind(&self, id: u16) -> Option<&(ModuleKind, M)>

Get module data by instance id, including ModuleKind

impl<M, State> ModuleRegistry<M, State>where M: Debug,

pub fn get_expect(&self, id: u16) -> &M

Return the module data belonging to the module identified by the supplied module_id

Panics

If the module isn’t in the registry

pub fn register_module(&mut self, id: u16, kind: ModuleKind, module: M)

Add a module to the registry

impl ModuleRegistry<Decoder, DecodingMode>

pub fn with_fallback(self) -> ModuleRegistry<Decoder, DecodingMode>

pub fn decoding_mode(&self) -> DecodingMode

pub fn assert_reject_mode(&self)

Panic if the Self::decoding_mode is not Reject