Trait fedimint_server::module::IModuleGen

pub trait IModuleGen: Debug {
    // Required methods
    fn decoder(&self) -> DynDecoder;
    fn versions(
        &self,
        core: CoreConsensusVersion
    ) -> Vec<ModuleConsensusVersion, Global> ;
    fn module_kind(&self) -> ModuleKind;
    fn init<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        cfg: ServerModuleConfig,
        db: Database,
        env: &'life1 BTreeMap<OsString, OsString, Global>,
        task_group: &'life2 mut TaskGroup
    ) -> Pin<Box<dyn Future<Output = Result<DynServerModule, Error>> + Send + 'async_trait, Global>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait,
             Self: 'async_trait;
    fn trusted_dealer_gen(
        &self,
        peers: &[PeerId],
        params: &ConfigGenParams
    ) -> BTreeMap<PeerId, ServerModuleConfig, Global>;
    fn distributed_gen<'life0, 'life1, 'life2, 'life3, 'life4, 'life5, 'async_trait>(
        &'life0 self,
        connections: &'life1 MuxPeerConnections<u16, DkgPeerMsg>,
        our_id: &'life2 PeerId,
        module_id: u16,
        peers: &'life3 [PeerId],
        params: &'life4 ConfigGenParams,
        task_group: &'life5 mut TaskGroup
    ) -> Pin<Box<dyn Future<Output = Result<Result<ServerModuleConfig, Cancelled>, Error>> + Send + 'async_trait, Global>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait,
             'life3: 'async_trait,
             'life4: 'async_trait,
             'life5: 'async_trait,
             Self: 'async_trait;
    fn to_config_response(
        &self,
        config: Value
    ) -> Result<ModuleConfigResponse, Error>;
    fn hash_client_module(&self, config: Value) -> Result<Hash, Error>;
    fn validate_config(
        &self,
        identity: &PeerId,
        config: ServerModuleConfig
    ) -> Result<(), Error>;
    fn dump_database<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        dbtx: &'life1 mut DatabaseTransaction<'life2>,
        prefix_names: Vec<String, Global>
    ) -> Pin<Box<dyn Future<Output = Box<dyn Iterator<Item = (String, Box<dyn Serialize + Send + 'static, Global>)> + 'life0, Global>> + Send + 'async_trait, Global>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait,
             Self: 'async_trait;
}

Interface for Module Generation

This trait contains the methods responsible for the module’s

initialization

config generation

config validation

Once the module configuration is ready, the module can be instantiated via [Self::init].

Required Methods

fn decoder(&self) -> DynDecoder

fn versions( &self, core: CoreConsensusVersion ) -> Vec<ModuleConsensusVersion, Global>

fn module_kind(&self) -> ModuleKind

fn init<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, cfg: ServerModuleConfig, db: Database, env: &'life1 BTreeMap<OsString, OsString, Global>, task_group: &'life2 mut TaskGroup ) -> Pin<Box<dyn Future<Output = Result<DynServerModule, Error>> + Send + 'async_trait, Global>>where 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait, Self: 'async_trait,

Initialize the DynServerModule instance from its config

fn trusted_dealer_gen( &self, peers: &[PeerId], params: &ConfigGenParams ) -> BTreeMap<PeerId, ServerModuleConfig, Global>

fn distributed_gen<'life0, 'life1, 'life2, 'life3, 'life4, 'life5, 'async_trait>( &'life0 self, connections: &'life1 MuxPeerConnections<u16, DkgPeerMsg>, our_id: &'life2 PeerId, module_id: u16, peers: &'life3 [PeerId], params: &'life4 ConfigGenParams, task_group: &'life5 mut TaskGroup ) -> Pin<Box<dyn Future<Output = Result<Result<ServerModuleConfig, Cancelled>, Error>> + Send + 'async_trait, Global>>where 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait, 'life3: 'async_trait, 'life4: 'async_trait, 'life5: 'async_trait, Self: 'async_trait,

fn to_config_response( &self, config: Value ) -> Result<ModuleConfigResponse, Error>

fn hash_client_module(&self, config: Value) -> Result<Hash, Error>

fn validate_config( &self, identity: &PeerId, config: ServerModuleConfig ) -> Result<(), Error>

fn dump_database<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, dbtx: &'life1 mut DatabaseTransaction<'life2>, prefix_names: Vec<String, Global> ) -> Pin<Box<dyn Future<Output = Box<dyn Iterator<Item = (String, Box<dyn Serialize + Send + 'static, Global>)> + 'life0, Global>> + Send + 'async_trait, Global>>where 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait, Self: 'async_trait,