Trait fedimint_server::module::IServerModuleInit

pub trait IServerModuleInit: IDynCommonModuleInit {
    // Required methods
    fn as_common(&self) -> &(dyn IDynCommonModuleInit + Sync + Send + 'static);
    fn supported_api_versions(&self) -> SupportedModuleApiVersions;
    fn database_version(&self) -> DatabaseVersion;
    fn init<'life0, 'life1, 'async_trait>(
        &'life0 self,
        cfg: ServerModuleConfig,
        db: Database,
        task_group: &'life1 mut TaskGroup
    ) -> Pin<Box<dyn Future<Output = Result<DynServerModule, Error>> + Send + 'async_trait, Global>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             Self: 'async_trait;
    fn get_database_migrations(
        &self
    ) -> BTreeMap<DatabaseVersion, fn(_: &'b mut DatabaseTransaction<'_>) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'b, Global>>, Global>;
    fn validate_params(
        &self,
        params: &ConfigGenModuleParams
    ) -> Result<(), Error>;
    fn trusted_dealer_gen(
        &self,
        peers: &[PeerId],
        params: &ConfigGenModuleParams
    ) -> BTreeMap<PeerId, ServerModuleConfig, Global>;
    fn distributed_gen<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        peers: &'life1 PeerHandle<'_>,
        params: &'life2 ConfigGenModuleParams
    ) -> Pin<Box<dyn Future<Output = Result<ServerModuleConfig, DkgError>> + Send + 'async_trait, Global>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait,
             Self: 'async_trait;
    fn validate_config(
        &self,
        identity: &PeerId,
        config: ServerModuleConfig
    ) -> Result<(), Error>;
    fn get_client_config(
        &self,
        module_instance_id: u16,
        config: &ServerModuleConsensusConfig
    ) -> Result<ClientModuleConfig, Error>;
}

Interface for Module Generation

This trait contains the methods responsible for the module’s

initialization

config generation

config validation

Once the module configuration is ready, the module can be instantiated via [Self::init].

Required Methods

fn as_common(&self) -> &(dyn IDynCommonModuleInit + Sync + Send + 'static)

fn supported_api_versions(&self) -> SupportedModuleApiVersions

fn database_version(&self) -> DatabaseVersion

fn init<'life0, 'life1, 'async_trait>( &'life0 self, cfg: ServerModuleConfig, db: Database, task_group: &'life1 mut TaskGroup ) -> Pin<Box<dyn Future<Output = Result<DynServerModule, Error>> + Send + 'async_trait, Global>>where 'life0: 'async_trait, 'life1: 'async_trait, Self: 'async_trait,

Initialize the DynServerModule instance from its config

fn get_database_migrations( &self ) -> BTreeMap<DatabaseVersion, fn(_: &'b mut DatabaseTransaction<'_>) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'b, Global>>, Global>

Retrieves the MigrationMap from the module to be applied to the database before the module is initialized. The MigrationMap is indexed on the from version.

fn validate_params(&self, params: &ConfigGenModuleParams) -> Result<(), Error>

fn trusted_dealer_gen( &self, peers: &[PeerId], params: &ConfigGenModuleParams ) -> BTreeMap<PeerId, ServerModuleConfig, Global>

fn distributed_gen<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, peers: &'life1 PeerHandle<'_>, params: &'life2 ConfigGenModuleParams ) -> Pin<Box<dyn Future<Output = Result<ServerModuleConfig, DkgError>> + Send + 'async_trait, Global>>where 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait, Self: 'async_trait,

fn validate_config( &self, identity: &PeerId, config: ServerModuleConfig ) -> Result<(), Error>

fn get_client_config( &self, module_instance_id: u16, config: &ServerModuleConsensusConfig ) -> Result<ClientModuleConfig, Error>