Trait fedimint_server::module::IntoModuleError

pub trait IntoModuleError {
    type Target;

    // Required method
    fn into_module_error_other(self) -> Self::Target;
}

Extension trait with a function to map Results used by modules to ModuleError

Currently each module defined it’s own enum XyzError { ... } and is not using anyhow::Error. For ? to work seamlessly two conversion would have to be made: enum-Error -> anyhow::Error -> enum-Error, while Into/From can only do one.

To avoid the boilerplate, this trait defines an easy conversion method.

Required Associated Types

type Target

Required Methods

fn into_module_error_other(self) -> Self::Target