Trait fedimint_server::module::ServerModule

pub trait ServerModule: Debug + Sized {
    type Common: ModuleCommon;
    type Gen: ServerModuleGen;
    type VerificationCache: VerificationCache;

    fn await_consensus_proposal<'a, 'life0, 'life1, 'async_trait>(
        &'a self,
        dbtx: &'life0 mut ModuleDatabaseTransaction<'life1, u16>
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait, Global>>
    where
        'a: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn consensus_proposal<'a, 'life0, 'life1, 'async_trait>(
        &'a self,
        dbtx: &'life0 mut ModuleDatabaseTransaction<'life1, u16>
    ) -> Pin<Box<dyn Future<Output = ConsensusProposal<<Self::Common as ModuleCommon>::ConsensusItem>> + Send + 'async_trait, Global>>
    where
        'a: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn process_consensus_item<'a, 'b, 'life0, 'async_trait>(
        &'a self,
        dbtx: &'life0 mut ModuleDatabaseTransaction<'b, u16>,
        consensus_item: <Self::Common as ModuleCommon>::ConsensusItem,
        peer_id: PeerId
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait, Global>>
    where
        'a: 'async_trait,
        'b: 'async_trait,
        'life0: 'async_trait,
        Self: 'async_trait;
    fn build_verification_cache<'a>(
        &'a self,
        inputs: impl Iterator<Item = &'a <Self::Common as ModuleCommon>::Input> + MaybeSend
    ) -> Self::VerificationCache;
    fn process_input<'a, 'b, 'c, 'life0, 'life1, 'async_trait>(
        &'a self,
        dbtx: &'life0 mut ModuleDatabaseTransaction<'c, u16>,
        input: &'b <Self::Common as ModuleCommon>::Input,
        verification_cache: &'life1 Self::VerificationCache
    ) -> Pin<Box<dyn Future<Output = Result<InputMeta, ModuleError>> + Send + 'async_trait, Global>>
    where
        'a: 'async_trait,
        'b: 'async_trait,
        'c: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn process_output<'a, 'b, 'life0, 'async_trait>(
        &'a self,
        dbtx: &'life0 mut ModuleDatabaseTransaction<'b, u16>,
        output: &'a <Self::Common as ModuleCommon>::Output,
        out_point: OutPoint
    ) -> Pin<Box<dyn Future<Output = Result<TransactionItemAmount, ModuleError>> + Send + 'async_trait, Global>>
    where
        'a: 'async_trait,
        'b: 'async_trait,
        'life0: 'async_trait,
        Self: 'async_trait;
    fn output_status<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        dbtx: &'life1 mut ModuleDatabaseTransaction<'life2, u16>,
        out_point: OutPoint
    ) -> Pin<Box<dyn Future<Output = Option<<Self::Common as ModuleCommon>::OutputOutcome>> + Send + 'async_trait, Global>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait;
    fn audit<'life0, 'life1, 'life2, 'life3, 'async_trait>(
        &'life0 self,
        dbtx: &'life1 mut ModuleDatabaseTransaction<'life2, u16>,
        audit: &'life3 mut Audit
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait, Global>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        'life3: 'async_trait,
        Self: 'async_trait;
    fn api_endpoints(&self) -> Vec<ApiEndpoint<Self>, Global> ;

    fn module_kind() -> ModuleKind { ... }
    fn decoder() -> Decoder { ... }
}

Required Associated Types

type Common: ModuleCommon

type Gen: ServerModuleGen

type VerificationCache: VerificationCache

Required Methods

fn await_consensus_proposal<'a, 'life0, 'life1, 'async_trait>(
&'a self,
dbtx: &'life0 mut ModuleDatabaseTransaction<'life1, u16>
) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait, Global>>where
'a: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Blocks until a new consensus_proposal is available.

fn consensus_proposal<'a, 'life0, 'life1, 'async_trait>(
&'a self,
dbtx: &'life0 mut ModuleDatabaseTransaction<'life1, u16>
) -> Pin<Box<dyn Future<Output = ConsensusProposal<<Self::Common as ModuleCommon>::ConsensusItem>> + Send + 'async_trait, Global>>where
'a: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

This module’s contribution to the next consensus proposal

fn process_consensus_item<'a, 'b, 'life0, 'async_trait>(
&'a self,
dbtx: &'life0 mut ModuleDatabaseTransaction<'b, u16>,
consensus_item: <Self::Common as ModuleCommon>::ConsensusItem,
peer_id: PeerId
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait, Global>>where
'a: 'async_trait,
'b: 'async_trait,
'life0: 'async_trait,
Self: 'async_trait,

This function is called once for every consensus item. The function returns an error if any only if the consensus item does not change our state and therefore may be safely discarded by the atomic broadcast.

fn build_verification_cache<'a>(
&'a self,
inputs: impl Iterator<Item = &'a <Self::Common as ModuleCommon>::Input> + MaybeSend
) -> Self::VerificationCache

Some modules may have slow to verify inputs that would block transaction processing. If the slow part of verification can be modeled as a pure function not involving any system state we can build a lookup table in a hyper-parallelized manner. This function is meant for constructing such lookup tables.

fn process_input<'a, 'b, 'c, 'life0, 'life1, 'async_trait>(
&'a self,
dbtx: &'life0 mut ModuleDatabaseTransaction<'c, u16>,
input: &'b <Self::Common as ModuleCommon>::Input,
verification_cache: &'life1 Self::VerificationCache
) -> Pin<Box<dyn Future<Output = Result<InputMeta, ModuleError>> + Send + 'async_trait, Global>>where
'a: 'async_trait,
'b: 'async_trait,
'c: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Try to spend a transaction input. On success all necessary updates will be part of the database transaction. On failure (e.g. double spend) the database transaction is rolled back and the operation will take no effect.

fn process_output<'a, 'b, 'life0, 'async_trait>(
&'a self,
dbtx: &'life0 mut ModuleDatabaseTransaction<'b, u16>,
output: &'a <Self::Common as ModuleCommon>::Output,
out_point: OutPoint
) -> Pin<Box<dyn Future<Output = Result<TransactionItemAmount, ModuleError>> + Send + 'async_trait, Global>>where
'a: 'async_trait,
'b: 'async_trait,
'life0: 'async_trait,
Self: 'async_trait,

Try to create an output (e.g. issue notes, peg-out BTC, …). On success all necessary updates to the database will be part of the database transaction. On failure (e.g. double spend) the database transaction is rolled back and the operation will take no effect.

The supplied out_point identifies the operation (e.g. a peg-out or note issuance) and can be used to retrieve its outcome later using output_status.

fn output_status<'life0, 'life1, 'life2, 'async_trait>(
&'life0 self,
dbtx: &'life1 mut ModuleDatabaseTransaction<'life2, u16>,
out_point: OutPoint
) -> Pin<Box<dyn Future<Output = Option<<Self::Common as ModuleCommon>::OutputOutcome>> + Send + 'async_trait, Global>>where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

Retrieve the current status of the output. Depending on the module this might contain data needed by the client to access funds or give an estimate of when funds will be available. Returns None if the output is unknown, NOT if it is just not ready yet.

fn audit<'life0, 'life1, 'life2, 'life3, 'async_trait>(
&'life0 self,
dbtx: &'life1 mut ModuleDatabaseTransaction<'life2, u16>,
audit: &'life3 mut Audit
) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait, Global>>where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
'life3: 'async_trait,
Self: 'async_trait,

Queries the database and returns all assets and liabilities of the module.

Summing over all modules, if liabilities > assets then an error has occurred in the database and consensus should halt.

fn api_endpoints(&self) -> Vec<ApiEndpoint<Self>, Global>

Returns a list of custom API endpoints defined by the module. These are made available both to users as well as to other modules. They thus should be deterministic, only dependant on their input and the current epoch.

Provided Methods

fn module_kind() -> ModuleKind

fn decoder() -> Decoder

Returns a decoder for the following associated types of this module:

Input

Output

OutputOutcome

ConsensusItem