Struct fedimint_server::modules::ln::LightningModule

pub struct LightningModule {
    pub(crate) cfg: LightningModuleConfig,
}

The lightning module implements an account system. It does not have the privacy guarantees of the e-cash mint module but instead allows for smart contracting. There exist three contract types that can be used to “lock” accounts:

Account: an account locked with a schnorr public key

Outgoing: an account locked with an HTLC-like contract allowing to incentivize an external Lightning node to make payments for the funder

Incoming: a contract type that represents the acquisition of a preimage belonging to a hash. Every incoming contract is preceded by an offer that specifies how much the seller is asking for the preimage to a particular hash. It also contains some threshold-encrypted data. Once the contract is funded the data is decrypted. If it is a valid preimage the contract’s funds are now accessible to the creator of the offer, if not they are accessible to the funder.

These three primitives allow to integrate the federation with the wider Lightning network through a centralized but untrusted (except for availability) Lightning gateway server.

Fields
cfg: LightningModuleConfig
Implementations

impl LightningModule

pub fn new(cfg: LightningModuleConfig) -> LightningModule

pub async fn get_offer(
&self,
dbtx: &mut DatabaseTransaction<'_>,
payment_hash: Hash
) -> impl Future<Output = Option<IncomingContractOffer>>

pub async fn get_offers(
&self,
dbtx: &mut DatabaseTransaction<'_>
) -> impl Future<Output = Vec<IncomingContractOffer, Global>>

pub async fn get_contract_account(
&self,
dbtx: &mut DatabaseTransaction<'_>,
contract_id: ContractId
) -> impl Future<Output = Option<ContractAccount>>

pub async fn list_gateways(
&self,
dbtx: &mut DatabaseTransaction<'_>
) -> impl Future<Output = Vec<LightningGateway, Global>>

pub async fn register_gateway(
&self,
dbtx: &mut DatabaseTransaction<'_>,
gateway: LightningGateway
) -> impl Future<Output = ()>