Crate fedimint_server::modules::wallet

Modules

common

config

db

keys

tweakable

txoproof

Structs

PegOut

PegOutFees

PegOutSignatureItem

PendingTransaction

A peg-out tx that is ready to be broadcast with a tweak for the change UTXO

RoundConsensus

RoundConsensusItem

SpendableUTXO

UnsignedTransaction

A PSBT that is awaiting enough signatures from the federation to becoming a PendingTransaction

UnzipWalletConsensusItem

Wallet

WalletConfigGenParams

WalletConfigGenerator

WalletInput

WalletOutput

WalletOutputOutcome

Contains the Bitcoin transaction id of the transaction created by the withdraw request

WalletVerificationCache

Enums

ProcessPegOutSigError

WalletConsensusItem

WalletError

Constants

CONFIRMATION_TARGET

Traits

IterUnzipWalletConsensusItem

Functions

broadcast_pending_tx

is_address_valid_for_network

run_broadcast_pending_tx

Type Definitions

PartialSig

PegInDescriptor