Trait fedimint_server::modules::wallet::tweakable::Contract

pub trait Contract {
    fn encode<W>(&self, writer: &mut W) -> Result<(), Error>
    where
        W: Write;
}

An object that can be used as a ricardian contract to tweak a key

Required Methods

fn encode<W>(&self, writer: &mut W) -> Result<(), Error>where
W: Write,

Serialize the contract in a deterministic way to be used as a tweak