fedimint_server::net::api

Module announcement

Structs

ApiAnnouncementKey

ApiAnnouncementPrefix

Functions

get_api_urls

Returns a list of all peers and their respective API URLs taking into account announcements overwriting the URLs contained in the original configuration.

insert_signed_api_announcement_if_not_present 🔒

Checks if we already have a signed API endpoint announcement for our own identity in the database and creates one if not.

start_api_announcement_service