fedimint_server::net
Module connect

Provides an abstract network connection interface and multiple implementations

Structs

TlsConfig

TlsTcpConnector

TCP connector with encryption and authentication

Traits

Connector

Allows to connect to peers and to listen for incoming connections. Connections are message based and should be authenticated and encrypted for production deployments.

Functions

dns_sanitize

Sanitizes name as valid domain name

parse_host_port

Parses the host and port from a url

Type Aliases

ConnectResult

ConnectionListener

DynConnector