fedimint_server::net::connect
Trait Connector

pub trait Connector<M>:
    Send
    + Sync
    + 'static {
    // Required methods
    fn peers(&self) -> Vec<PeerId>;
    fn connect<'life0, 'async_trait>(
        &'life0 self,
        peer: PeerId,
    ) -> Pin<Box<dyn Future<Output = ConnectResult<M>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn listen<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<ConnectionListener<M>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;

    // Provided method
    fn into_dyn(self) -> DynConnector<M>
       where Self: Sized { ... }
}

Allows to connect to peers and to listen for incoming connections. Connections are message based and should be authenticated and encrypted for production deployments.

Required Methods

fn peers(&self) -> Vec<PeerId>

fn connect<'life0, 'async_trait>( &'life0 self, peer: PeerId, ) -> Pin<Box<dyn Future<Output = ConnectResult<M>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn listen<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<ConnectionListener<M>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Provided Methods

fn into_dyn(self) -> DynConnector<M>
where Self: Sized,