Struct fedimint_server::FedimintServer

pub struct FedimintServer {
    pub consensus: Arc<FedimintConsensus>,
    pub connections: PeerConnections<EpochMessage>,
    pub cfg: ServerConfig,
    pub hbbft: HoneyBadger<Vec<SerdeConsensusItem>, PeerId>,
    pub api: DynFederationApi,
    pub peers: BTreeSet<PeerId>,
    pub rejoin_at_epoch: Option<HashMap<u64, HashSet<PeerId>>>,
    pub run_empty_epochs: u64,
    pub last_processed_epoch: Option<SignedEpochOutcome>,
    pub decoders: ModuleDecoderRegistry,
}

Fields
consensus: Arc<FedimintConsensus>connections: PeerConnections<EpochMessage>cfg: ServerConfighbbft: HoneyBadger<Vec<SerdeConsensusItem>, PeerId>api: DynFederationApipeers: BTreeSet<PeerId>rejoin_at_epoch: Option<HashMap<u64, HashSet<PeerId>>>run_empty_epochs: u64last_processed_epoch: Option<SignedEpochOutcome>decoders: ModuleDecoderRegistry
Implementations

impl FedimintServer

pub async fn run(
cfg: ServerConfig,
consensus: FedimintConsensus,
decoders: ModuleDecoderRegistry,
task_group: &mut TaskGroup
) -> Result<()>

Start all the components of the mint and plug them together

pub async fn new(
cfg: ServerConfig,
consensus: FedimintConsensus,
decoders: ModuleDecoderRegistry,
task_group: &mut TaskGroup
) -> Self

pub async fn new_with(
cfg: ServerConfig,
consensus: FedimintConsensus,
connector: PeerConnector<EpochMessage>,
decoders: ModuleDecoderRegistry,
task_group: &mut TaskGroup
) -> Self

pub(crate) async fn run_consensus(self, task_handle: TaskHandle)

Loop run_conensus_epoch until shut down

pub async fn start_consensus(&mut self)

Starts consensus by skipping to the last saved epoch history and triggering a new epoch

pub(crate) fn next_epoch_to_process(&self) -> u64

Returns the next epoch that we need to process, based on our saved history

pub async fn process_outcome(
&mut self,
last_outcome: HbbftConsensusOutcome
) -> Result<(), EpochVerifyError>

Requests, verifies and processes history from peers

last_outcome - The consensus outcome (unprocessed), we’re trying to process.

pub async fn run_consensus_epoch(
&mut self,
proposal: impl Future<Output = ConsensusProposal>,
rng: &mut impl RngCore + CryptoRng + Clone + 'static
) -> Cancellable<Vec<HbbftConsensusOutcome>>

The main consensus function:

Await a new proposal event or receiving a proposal from peers

Send the ConsensusProposal to peers

Run HBBFT until a ConsensusOutcome can be returned

pub(crate) async fn handle_step(
&mut self,
step: Step<Vec<SerdeConsensusItem>, PeerId>
) -> Cancellable<Vec<HbbftConsensusOutcome>>

Handles one step of the HBBFT algorithm, sending messages to peers and parsing any outcomes contained in the step

pub(crate) async fn propose_epoch(
&mut self,
proposal: ConsensusProposal,
rng: &mut impl RngCore + CryptoRng + Clone + 'static
) -> Cancellable<Step<Vec<SerdeConsensusItem>, PeerId>>

pub(crate) async fn await_next_epoch(
&mut self
) -> Cancellable<Option<(PeerId, EpochMessage)>>

pub(crate) fn start_next_epoch(&self, msg: &(PeerId, EpochMessage)) -> bool

pub(crate) async fn handle_message(
&mut self,
msg: (PeerId, EpochMessage)
) -> Cancellable<Vec<HbbftConsensusOutcome>>

Runs a single HBBFT consensus step

pub(crate) async fn rejoin_at_epoch(&mut self, epoch: u64, peer: PeerId)

If we are rejoining and received a threshold of messages from the same epoch, then skip to that epoch. Give ourselves a buffer of NUM_EPOCHS_REJOIN_AHEAD so we can ensure we receive enough HBBFT messages to produce an outcome.

pub(crate) async fn request_rejoin(&mut self, epochs_to_run: u64)

Sends a rejoin request to all peers, indicating the number of epochs we want them to create