Trait fedimint_server::transaction::TransactionItem

pub trait TransactionItem {
    fn amount(&self) -> Amount;
    fn fee(&self, fee_consensus: &FeeConsensus) -> Amount;
}

Common properties of value transfer for inputs and outputs.

Input and output both define amount transferred and fee. The fee is paid to the federation.

Required Methods

fn amount(&self) -> Amount

The amount before fees represented by the in/output

fn fee(&self, fee_consensus: &FeeConsensus) -> Amount

The fee that will be charged for this in/output