Struct fedimint_server::api::WsFederationApi

pub struct WsFederationApi<C = Client> {
    peers: BTreeSet<PeerId, Global>,
    members: Vec<FederationMember<C>, Global>,
}

Mint API client that will try to run queries against all members expecting equal results from at least min_eq_results of them. Members that return differing results are returned as a member faults list.

Fields
peers: BTreeSet<PeerId, Global>members: Vec<FederationMember<C>, Global>
Implementations

impl WsFederationApi<Client>

pub fn new(members: Vec<(PeerId, Url), Global>) -> WsFederationApi<Client>

Creates a new API client

pub fn from_config(config: &ClientConfig) -> WsFederationApi<Client>

Creates a new API client from a client config

pub fn from_connect_info(
info: &[WsClientConnectInfo]
) -> WsFederationApi<Client>

Creates a new API client from a connect info, assumes they are in peer id order

impl<C> WsFederationApi<C>

pub fn peers(&self) -> Vec<PeerId, Global>

pub fn new_with_client(
members: Vec<(PeerId, Url), Global>
) -> WsFederationApi<C>

Creates a new API client