Trait fedimint_server::api::GlobalFederationApi

pub trait GlobalFederationApi {
    // Required methods
    fn submit_transaction<'life0, 'async_trait>(
        &'life0 self,
        tx: Transaction
    ) -> Pin<Box<dyn Future<Output = Result<TransactionId, FederationError>> + Send + 'async_trait, Global>>
       where 'life0: 'async_trait,
             Self: 'async_trait;
    fn fetch_tx_outcome<'life0, 'life1, 'async_trait>(
        &'life0 self,
        txid: &'life1 TransactionId
    ) -> Pin<Box<dyn Future<Output = Result<Option<TransactionStatus>, FederationError>> + Send + 'async_trait, Global>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             Self: 'async_trait;
    fn await_tx_outcome<'life0, 'life1, 'async_trait>(
        &'life0 self,
        txid: &'life1 TransactionId
    ) -> Pin<Box<dyn Future<Output = Result<TransactionStatus, FederationError>> + Send + 'async_trait, Global>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             Self: 'async_trait;
    fn fetch_epoch_history<'life0, 'life1, 'async_trait>(
        &'life0 self,
        epoch: u64,
        epoch_pk: PublicKey,
        decoders: &'life1 ModuleRegistry<Decoder, DecodingMode>
    ) -> Pin<Box<dyn Future<Output = Result<SignedEpochOutcome, FederationError>> + Send + 'async_trait, Global>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             Self: 'async_trait;
    fn fetch_epoch_count<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<u64, FederationError>> + Send + 'async_trait, Global>>
       where 'life0: 'async_trait,
             Self: 'async_trait;
    fn fetch_output_outcome<'life0, 'life1, 'async_trait, R>(
        &'life0 self,
        out_point: OutPoint,
        module_decoder: &'life1 Decoder
    ) -> Pin<Box<dyn Future<Output = Result<Option<R>, OutputOutcomeError>> + Send + 'async_trait, Global>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             R: OutputOutcome + 'async_trait,
             Self: 'async_trait;
    fn await_output_outcome<'life0, 'life1, 'async_trait, R>(
        &'life0 self,
        outpoint: OutPoint,
        timeout: Duration,
        module_decoder: &'life1 Decoder
    ) -> Pin<Box<dyn Future<Output = Result<R, OutputOutcomeError>> + Send + 'async_trait, Global>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             R: OutputOutcome + 'async_trait,
             Self: 'async_trait;
    fn download_client_config<'life0, 'life1, 'async_trait>(
        &'life0 self,
        info: &'life1 InviteCode
    ) -> Pin<Box<dyn Future<Output = Result<ClientConfig, FederationError>> + Send + 'async_trait, Global>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             Self: 'async_trait;
    fn consensus_config_hash<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Hash, FederationError>> + Send + 'async_trait, Global>>
       where 'life0: 'async_trait,
             Self: 'async_trait;
    fn upload_backup<'life0, 'life1, 'async_trait>(
        &'life0 self,
        request: &'life1 SignedBackupRequest
    ) -> Pin<Box<dyn Future<Output = Result<(), FederationError>> + Send + 'async_trait, Global>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             Self: 'async_trait;
    fn download_backup<'life0, 'life1, 'async_trait>(
        &'life0 self,
        id: &'life1 XOnlyPublicKey
    ) -> Pin<Box<dyn Future<Output = Result<Vec<ClientBackupSnapshot, Global>, FederationError>> + Send + 'async_trait, Global>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             Self: 'async_trait;
    fn discover_api_version_set<'life0, 'life1, 'async_trait>(
        &'life0 self,
        client_versions: &'life1 SupportedApiVersionsSummary
    ) -> Pin<Box<dyn Future<Output = Result<ApiVersionSet, FederationError>> + Send + 'async_trait, Global>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             Self: 'async_trait;
}

The API for the global (non-module) endpoints

Required Methods

fn submit_transaction<'life0, 'async_trait>( &'life0 self, tx: Transaction ) -> Pin<Box<dyn Future<Output = Result<TransactionId, FederationError>> + Send + 'async_trait, Global>>where 'life0: 'async_trait, Self: 'async_trait,

fn fetch_tx_outcome<'life0, 'life1, 'async_trait>( &'life0 self, txid: &'life1 TransactionId ) -> Pin<Box<dyn Future<Output = Result<Option<TransactionStatus>, FederationError>> + Send + 'async_trait, Global>>where 'life0: 'async_trait, 'life1: 'async_trait, Self: 'async_trait,

fn await_tx_outcome<'life0, 'life1, 'async_trait>( &'life0 self, txid: &'life1 TransactionId ) -> Pin<Box<dyn Future<Output = Result<TransactionStatus, FederationError>> + Send + 'async_trait, Global>>where 'life0: 'async_trait, 'life1: 'async_trait, Self: 'async_trait,

fn fetch_epoch_history<'life0, 'life1, 'async_trait>( &'life0 self, epoch: u64, epoch_pk: PublicKey, decoders: &'life1 ModuleRegistry<Decoder, DecodingMode> ) -> Pin<Box<dyn Future<Output = Result<SignedEpochOutcome, FederationError>> + Send + 'async_trait, Global>>where 'life0: 'async_trait, 'life1: 'async_trait, Self: 'async_trait,

fn fetch_epoch_count<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<u64, FederationError>> + Send + 'async_trait, Global>>where 'life0: 'async_trait, Self: 'async_trait,

fn fetch_output_outcome<'life0, 'life1, 'async_trait, R>( &'life0 self, out_point: OutPoint, module_decoder: &'life1 Decoder ) -> Pin<Box<dyn Future<Output = Result<Option<R>, OutputOutcomeError>> + Send + 'async_trait, Global>>where 'life0: 'async_trait, 'life1: 'async_trait, R: OutputOutcome + 'async_trait, Self: 'async_trait,

fn await_output_outcome<'life0, 'life1, 'async_trait, R>( &'life0 self, outpoint: OutPoint, timeout: Duration, module_decoder: &'life1 Decoder ) -> Pin<Box<dyn Future<Output = Result<R, OutputOutcomeError>> + Send + 'async_trait, Global>>where 'life0: 'async_trait, 'life1: 'async_trait, R: OutputOutcome + 'async_trait, Self: 'async_trait,

fn download_client_config<'life0, 'life1, 'async_trait>( &'life0 self, info: &'life1 InviteCode ) -> Pin<Box<dyn Future<Output = Result<ClientConfig, FederationError>> + Send + 'async_trait, Global>>where 'life0: 'async_trait, 'life1: 'async_trait, Self: 'async_trait,

Fetch client configuration info only if verified against a federation id

fn consensus_config_hash<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<Hash, FederationError>> + Send + 'async_trait, Global>>where 'life0: 'async_trait, Self: 'async_trait,

Fetches the server consensus hash if enough peers agree on it

fn upload_backup<'life0, 'life1, 'async_trait>( &'life0 self, request: &'life1 SignedBackupRequest ) -> Pin<Box<dyn Future<Output = Result<(), FederationError>> + Send + 'async_trait, Global>>where 'life0: 'async_trait, 'life1: 'async_trait, Self: 'async_trait,

fn download_backup<'life0, 'life1, 'async_trait>( &'life0 self, id: &'life1 XOnlyPublicKey ) -> Pin<Box<dyn Future<Output = Result<Vec<ClientBackupSnapshot, Global>, FederationError>> + Send + 'async_trait, Global>>where 'life0: 'async_trait, 'life1: 'async_trait, Self: 'async_trait,

fn discover_api_version_set<'life0, 'life1, 'async_trait>( &'life0 self, client_versions: &'life1 SupportedApiVersionsSummary ) -> Pin<Box<dyn Future<Output = Result<ApiVersionSet, FederationError>> + Send + 'async_trait, Global>>where 'life0: 'async_trait, 'life1: 'async_trait, Self: 'async_trait,

Query peers and calculate optimal common api versions to use.