Module fedimint_server::config

Fedimint toplevel config

Modules

api

distributedgen

io

serde_binary_human_readable

serde_tls_cert_map 🔒

serde_tls_key 🔒

Structs

ClientConfig

Total client config

ConfigGenParams

All the parameters necessary for generating the ServerConfig during setup

FederationId

The federation id is a copy of the authentication threshold public key of the federation

GlobalClientConfig

Federation-wide client config

JsonWithKind

serde_json::Value that must contain kind: String field

ModuleInitRegistry

PeerUrl

ServerConfig

All the serializable configuration for the fedimint server

ServerConfigConsensus

ServerConfigLocal

ServerConfigPrivate

ServerModuleConfig

Config for the server-side of a particular Federation module

ServerModuleConsensusConfig

Enums

DkgError

Captures an error occurring in DKG

DkgPeerMsg

Things that a distributed_gen config can send between peers

KeyType

The types of keys to run distributed key generation for

Constants

CORE_CONSENSUS_VERSION

DEFAULT_BROADCAST_EXPECTED_ROUNDS_PER_SESSION 🔒

DEFAULT_BROADCAST_MAX_ROUNDS_PER_SESSION 🔒

DEFAULT_BROADCAST_ROUND_DELAY_MS 🔒

DEFAULT_MAX_CLIENT_CONNECTIONS 🔒

The default maximum open connections the API can handle

ENV_MAX_CLIENT_CONNECTIONS 🔒

The env var for maximum open connections the API can handle

Traits

TypedServerModuleConfig

Module (server side) config, typed

Functions

connect

gen_cert_and_key

max_connections

Type Definitions

DkgResult

Result of running DKG

ServerModuleInitRegistry