Module fedimint_server::config

Fedimint toplevel config

Modules

api

distributedgen

io

serde_binary_human_readable

serde_tls_cert_map 🔒

serde_tls_key 🔒

Structs

ClientConfig

Total client config

ClientConfigResponse

The API response for client config requests, signed by the Federation

ClientModuleConfig

Config for the client-side of a particular Federation module

ConfigGenModuleParams

Type erased ModuleInitParams used to generate the ServerModuleConfig during config gen

ConfigGenParams

All the parameters necessary for generating the ServerConfig during setup

EmptyGenParams

Empty struct for if there are no params

FederationId

The federation id is a copy of the authentication threshold public key of the federation

FederationIdPrefix

Prefix of the FederationId, useful for UX improvements

GlobalClientConfig

Federation-wide client config

JsonClientConfig

Client config that cannot be cryptographically verified but is easier to parse by external tools

JsonWithKind

serde_json::Value that must contain kind: String field

ModuleInitRegistry

PeerUrl

ServerConfig

All the serializable configuration for the fedimint server

ServerConfigConsensus

ServerConfigLocal

ServerConfigPrivate

ServerModuleConfig

Config for the server-side of a particular Federation module

ServerModuleConsensusConfig

Enums

DkgError

Captures an error occurring in DKG

DkgMessage

DkgPeerMsg

Things that a distributed_gen config can send between peers

KeyType

The types of keys to run distributed key generation for

SupportedDkgMessage

enum version of SupportedDkgMessage

Constants

CORE_CONSENSUS_VERSION

DEFAULT_MAX_CLIENT_CONNECTIONS 🔒

The default maximum open connections the API can handle

ENV_MAX_CLIENT_CONNECTIONS 🔒

The env var for maximum open connections the API can handle

META_FEDERATION_NAME_KEY

Key under which the federation name can be sent to client in the meta part of the config

Traits

DkgGroup

Defines a group (e.g. G1 or G2) that we can generate keys for

ISupportedDkgMessage

Supported (by Fedimint’s code) DkgMessage<T> types

ModuleInitParams

SGroup

TypedServerModuleConfig

Module (server side) config, typed

TypedServerModuleConsensusConfig

Consensus-critical part of a server side module config

Functions

connect

gen_cert_and_key

load_from_file

max_connections

Type Aliases

CommonModuleInitRegistry

DkgResult

Result of running DKG

ServerModuleConfigGenParamsRegistry

Registry that contains the config gen params for all modules

ServerModuleInitRegistry