Module fedimint_server::config

Fedimint toplevel config

Modules

api

distributedgen

io

serde_binary_human_readable

serde_tls_cert_map 🔒

serde_tls_key 🔒

Structs

ClientConfig

Total client config

ClientModuleConfig

Config for the client-side of a particular Federation module

ConfigGenModuleParams

Module generation (so passed to dkg, not to the module itself) config parameters

ConfigGenParams

All the parameters necessary for generating the ServerConfig during setup

ConfigResponse

The API response for configuration requests

FederationId

The federation id is a copy of the authentication threshold public key of the federation

JsonWithKind

serde_json::Value that must contain kind: String field

LegacyInitOrderIter

Iterate over module generators in a legacy, hardcoded order: ln, mint, wallet, rest… Returning each kind exactly once, so that LEGACY_HARDCODED_ constants correspond to correct module kind.

ModuleConfigResponse

Response from the API for this particular module

ModuleGenRegistry

PeerUrl

ServerConfig

All the serializable configuration for the fedimint server

ServerConfigConsensus

ServerConfigLocal

ServerConfigPrivate

ServerModuleConfig

Config for the server-side of a particular Federation module

Enums

DkgError

Captures an error occurring in DKG

DkgMessage

DkgPeerMsg

Things that a distributed_gen config can send between peers

KeyType

The types of keys to run distributed key generation for

SupportedDkgMessage

enum version of SupportedDkgMessage

Constants

CORE_CONSENSUS_VERSION

DEFAULT_CONFIG_DOWNLOAD_LIMIT 🔒

How many times a config download token can be used by a client

DEFAULT_MAX_CLIENT_CONNECTIONS 🔒

The maximum open connections the API can handle

META_FEDERATION_NAME_KEY

Key under which the federation name can be sent to client in the meta part of the config

Traits

DkgGroup

Defines a group (e.g. G1 or G2) that we can generate keys for

ISupportedDkgMessage

Supported (by Fedimint’s code) DkgMessage<T> types

ModuleGenParams

SGroup

TypedClientModuleConfig

Typed client side module config

TypedServerModuleConfig

Module (server side) config

TypedServerModuleConsensusConfig

Consensus-critical part of a server side module config

Functions

connect

gen_cert_and_key

load_from_file

Type Definitions

CommonModuleGenRegistry

DkgResult

Result of running DKG

ServerModuleGenParamsRegistry

Configs for each module’s DKG

ServerModuleGenRegistry