fedimint_server::config
Module api

Structs

ConfigGenApi

Serves the config gen API endpoints

ConfigGenLocalConnection 🔒

Our local connection info

ConfigGenParamsLocal

Config gen params that are only used locally, shouldn’t be shared

ConfigGenSettings

All the info we configure prior to config gen starting

ConfigGenState

State held by the API after receiving a ConfigGenConnectionsRequest

Enums

BitcoinRpcConnectionStatus

Functions

check_bitcoind_status 🔒

server_endpoints