Module fedimint_server::config::io

Constants

CLIENT_CONFIG

Client configuration file

CLIENT_INVITE_CODE_FILE

Client connection string file

CONFIG_STAGING_DIR

Temporary directiry where server configs are stored / removed through the setup process On setup complete, the configs are moved to the server config directory and the staging directory is removed

CONSENSUS_CONFIG

Server consensus-only configurable file

DB_FILE

Database file name

ENCRYPTED_EXT

JSON_EXT

LOCAL_CONFIG

Server locally configurable file

PLAINTEXT_PASSWORD

Plain-text stored password, used to restart the server without having to send a password in via the API

PRIVATE_CONFIG

Server encrypted private keys file

SALT_FILE

Salt backup for combining with the private key

Functions

encrypted_json_read 🔒

Reads an encrypted json file into a struct

encrypted_json_write 🔒

Writes struct into an encrypted json file

plaintext_display_write 🔒

plaintext_json_read 🔒

Reads a plaintext json file into a struct

plaintext_json_write 🔒

Writes struct into a plaintext json file

read_server_config

Reads the server from the local, private, and consensus cfg files

write_server_config

Writes the server into configuration files (private keys encrypted)