fedimint_server::consensus::api

Struct ConsensusApi

pub struct ConsensusApi {

    pub cfg: ServerConfig,
    pub cfg_dir: PathBuf,
    pub db: Database,
    pub modules: ServerModuleRegistry,
    pub client_cfg: ClientConfig,
    pub force_api_secret: Option<String>,
    pub submission_sender: Sender<ConsensusItem>,
    pub shutdown_receiver: Receiver<Option<u64>>,
    pub shutdown_sender: Sender<Option<u64>>,
    pub ord_latency_receiver: Receiver<Option<Duration>>,
    pub p2p_status_receivers: P2PStatusReceivers,
    pub ci_status_receivers: BTreeMap<PeerId, Receiver<Option<u64>>>,
    pub bitcoin_rpc_connection: ServerBitcoinRpcMonitor,
    pub supported_api_versions: SupportedApiVersionsSummary,
    pub code_version_str: String,
    pub task_group: TaskGroup,

}

Fields
cfg: ServerConfig

Our server configuration

cfg_dir: PathBuf

Directory where config files are stored

db: Database

Database for serving the API

modules: ServerModuleRegistry

Modules registered with the federation

client_cfg: ClientConfig

Cached client config

force_api_secret: Option<String>submission_sender: Sender<ConsensusItem>

For sending API events to consensus such as transactions

shutdown_receiver: Receiver<Option<u64>>shutdown_sender: Sender<Option<u64>>ord_latency_receiver: Receiver<Option<Duration>>p2p_status_receivers: P2PStatusReceiversci_status_receivers: BTreeMap<PeerId, Receiver<Option<u64>>>bitcoin_rpc_connection: ServerBitcoinRpcMonitorsupported_api_versions: SupportedApiVersionsSummarycode_version_str: Stringtask_group: TaskGroup
Implementations

impl ConsensusApi

pub fn api_versions_summary(&self) -> &SupportedApiVersionsSummary

pub fn get_active_api_secret(&self) -> Option<String>

pub async fn submit_transaction( &self, transaction: Transaction, ) -> Result<TransactionId, TransactionError>

pub async fn await_transaction( &self, txid: TransactionId, ) -> (Vec<ModuleInstanceId>, DatabaseTransaction<'_, Committable>)

pub async fn await_output_outcome( &self, outpoint: OutPoint, ) -> Result<SerdeModuleEncoding<DynOutputOutcome>>

pub async fn await_outputs_outcomes( &self, outpoint_range: OutPointRange, ) -> Result<Vec<Option<SerdeModuleEncoding<DynOutputOutcome>>>>

pub async fn session_count(&self) -> u64

pub async fn await_signed_session_outcome( &self, index: u64, ) -> SignedSessionOutcome

pub async fn session_status(&self, session_index: u64) -> SessionStatusV2

pub async fn get_federation_status(&self) -> ApiResult<LegacyFederationStatus>

fn shutdown(&self, index: Option<u64>)

async fn get_federation_audit(&self) -> ApiResult<AuditSummary>

fn get_guardian_config_backup( &self, password: &str, _auth: &GuardianAuthToken, ) -> GuardianConfigBackup

Uses the in-memory config to write a config backup tar archive that guardians can download. Private keys are encrypted with the guardian password, so it should be safe to store anywhere, this also means the backup is useless without the password.

async fn handle_backup_request( &self, dbtx: &mut DatabaseTransaction<'_>, request: SignedBackupRequest, ) -> Result<(), ApiError>

async fn handle_recover_request( &self, dbtx: &mut DatabaseTransaction<'_>, id: PublicKey, ) -> Option<ClientBackupSnapshot>

async fn api_announcements(&self) -> BTreeMap<PeerId, SignedApiAnnouncement>

List API URL announcements from all peers we have received them from (at least ourselves)

fn fedimintd_version(&self) -> String

Returns the tagged fedimintd version currently running

async fn submit_api_announcement( &self, peer_id: PeerId, announcement: SignedApiAnnouncement, ) -> Result<(), ApiError>

Add an API URL announcement from a peer to our database to be returned by ConsensusApi::api_announcements.

async fn sign_api_announcement(&self, new_url: SafeUrl) -> SignedApiAnnouncement

fn change_guardian_password( &self, new_password: &str, _auth: &GuardianAuthToken, ) -> Result<(), ApiError>

Changes the guardian password by re-encrypting the private config and changing the on-disk password file if present. fedimintd is shut down afterward, the user’s service manager (e.g. systemd is expected to restart it).