Trait fedimint_server::core::server::IServerModule

pub trait IServerModule: Debug {
    // Required methods
    fn as_any(&self) -> &(dyn Any + 'static);
    fn decoder(&self) -> Decoder;
    fn await_consensus_proposal<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        dbtx: &'life1 mut ModuleDatabaseTransaction<'life2, u16>
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait, Global>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait,
             Self: 'async_trait;
    fn consensus_proposal<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        dbtx: &'life1 mut ModuleDatabaseTransaction<'life2, u16>,
        module_instance_id: u16
    ) -> Pin<Box<dyn Future<Output = ConsensusProposal<DynModuleConsensusItem>> + Send + 'async_trait, Global>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait,
             Self: 'async_trait;
    fn process_consensus_item<'a, 'life0, 'life1, 'async_trait>(
        &'life0 self,
        dbtx: &'life1 mut ModuleDatabaseTransaction<'a, u16>,
        consensus_item: DynModuleConsensusItem,
        peer_id: PeerId
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait, Global>>
       where 'a: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             Self: 'async_trait;
    fn build_verification_cache(
        &self,
        inputs: &[DynInput]
    ) -> DynVerificationCache;
    fn process_input<'a, 'b, 'c, 'life0, 'life1, 'async_trait>(
        &'a self,
        dbtx: &'life0 mut ModuleDatabaseTransaction<'c, u16>,
        input: &'b DynInput,
        verification_cache: &'life1 DynVerificationCache
    ) -> Pin<Box<dyn Future<Output = Result<InputMeta, ModuleError>> + Send + 'async_trait, Global>>
       where 'a: 'async_trait,
             'b: 'async_trait,
             'c: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             Self: 'async_trait;
    fn process_output<'a, 'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        dbtx: &'life1 mut ModuleDatabaseTransaction<'a, u16>,
        output: &'life2 DynOutput,
        out_point: OutPoint
    ) -> Pin<Box<dyn Future<Output = Result<TransactionItemAmount, ModuleError>> + Send + 'async_trait, Global>>
       where 'a: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait,
             Self: 'async_trait;
    fn output_status<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        dbtx: &'life1 mut ModuleDatabaseTransaction<'life2, u16>,
        out_point: OutPoint,
        module_instance_id: u16
    ) -> Pin<Box<dyn Future<Output = Option<DynOutputOutcome>> + Send + 'async_trait, Global>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait,
             Self: 'async_trait;
    fn audit<'life0, 'life1, 'life2, 'life3, 'async_trait>(
        &'life0 self,
        dbtx: &'life1 mut ModuleDatabaseTransaction<'life2, u16>,
        audit: &'life3 mut Audit,
        module_instance_id: u16
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait, Global>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait,
             'life3: 'async_trait,
             Self: 'async_trait;
    fn api_endpoints(&self) -> Vec<ApiEndpoint<DynServerModule>, Global>;
}

Backend side module interface

Server side Fedimint module needs to implement this trait.

Required Methods

fn as_any(&self) -> &(dyn Any + 'static)

fn decoder(&self) -> Decoder

Returns the decoder belonging to the server module

fn await_consensus_proposal<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, dbtx: &'life1 mut ModuleDatabaseTransaction<'life2, u16> ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait, Global>>where 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait, Self: 'async_trait,

Blocks until a new consensus_proposal is available.

fn consensus_proposal<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, dbtx: &'life1 mut ModuleDatabaseTransaction<'life2, u16>, module_instance_id: u16 ) -> Pin<Box<dyn Future<Output = ConsensusProposal<DynModuleConsensusItem>> + Send + 'async_trait, Global>>where 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait, Self: 'async_trait,

This module’s contribution to the next consensus proposal

fn process_consensus_item<'a, 'life0, 'life1, 'async_trait>( &'life0 self, dbtx: &'life1 mut ModuleDatabaseTransaction<'a, u16>, consensus_item: DynModuleConsensusItem, peer_id: PeerId ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait, Global>>where 'a: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, Self: 'async_trait,

This function is called once for every consensus item. The function returns an error if any only if the consensus item does not change our state and therefore may be safely discarded by the atomic broadcast.

fn build_verification_cache(&self, inputs: &[DynInput]) -> DynVerificationCache

Some modules may have slow to verify inputs that would block transaction processing. If the slow part of verification can be modeled as a pure function not involving any system state we can build a lookup table in a hyper-parallelized manner. This function is meant for constructing such lookup tables.

fn process_input<'a, 'b, 'c, 'life0, 'life1, 'async_trait>( &'a self, dbtx: &'life0 mut ModuleDatabaseTransaction<'c, u16>, input: &'b DynInput, verification_cache: &'life1 DynVerificationCache ) -> Pin<Box<dyn Future<Output = Result<InputMeta, ModuleError>> + Send + 'async_trait, Global>>where 'a: 'async_trait, 'b: 'async_trait, 'c: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, Self: 'async_trait,

Try to spend a transaction input. On success all necessary updates will be part of the database transaction. On failure (e.g. double spend) the database transaction is rolled back and the operation will take no effect.

fn process_output<'a, 'life0, 'life1, 'life2, 'async_trait>( &'life0 self, dbtx: &'life1 mut ModuleDatabaseTransaction<'a, u16>, output: &'life2 DynOutput, out_point: OutPoint ) -> Pin<Box<dyn Future<Output = Result<TransactionItemAmount, ModuleError>> + Send + 'async_trait, Global>>where 'a: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait, Self: 'async_trait,

Try to create an output (e.g. issue notes, peg-out BTC, …). On success all necessary updates to the database will be part of the database transaction. On failure (e.g. double spend) the database transaction is rolled back and the operation will take no effect.

The supplied out_point identifies the operation (e.g. a peg-out or note issuance) and can be used to retrieve its outcome later using output_status.

fn output_status<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, dbtx: &'life1 mut ModuleDatabaseTransaction<'life2, u16>, out_point: OutPoint, module_instance_id: u16 ) -> Pin<Box<dyn Future<Output = Option<DynOutputOutcome>> + Send + 'async_trait, Global>>where 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait, Self: 'async_trait,

Retrieve the current status of the output. Depending on the module this might contain data needed by the client to access funds or give an estimate of when funds will be available. Returns None if the output is unknown, NOT if it is just not ready yet.

fn audit<'life0, 'life1, 'life2, 'life3, 'async_trait>( &'life0 self, dbtx: &'life1 mut ModuleDatabaseTransaction<'life2, u16>, audit: &'life3 mut Audit, module_instance_id: u16 ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait, Global>>where 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait, 'life3: 'async_trait, Self: 'async_trait,

Queries the database and returns all assets and liabilities of the module.

Summing over all modules, if liabilities > assets then an error has occurred in the database and consensus should halt.

fn api_endpoints(&self) -> Vec<ApiEndpoint<DynServerModule>, Global>

Returns a list of custom API endpoints defined by the module. These are made available both to users as well as to other modules. They thus should be deterministic, only dependant on their input and the current epoch.