Trait fedimint_server::encoding::Encodable

pub trait Encodable {
    fn consensus_encode<W>(&self, writer: &mut W) -> Result<usize, Error>
    where
        W: Write;

    fn consensus_encode_to_vec(&self) -> Result<Vec<u8, Global>, Error> { ... }
    fn consensus_encode_to_hex(&self) -> Result<String, Error> { ... }
    fn consensus_hash(&self) -> Hash { ... }
}

Data which can be encoded in a consensus-consistent way

Required Methods

fn consensus_encode<W>(&self, writer: &mut W) -> Result<usize, Error>where
W: Write,

Encode an object with a well-defined format. Returns the number of bytes written on success.

The only errors returned are errors propagated from the writer.

Provided Methods

fn consensus_encode_to_vec(&self) -> Result<Vec<u8, Global>, Error>

Self::consensus_encode to newly allocated Vec<u8>

fn consensus_encode_to_hex(&self) -> Result<String, Error>

fn consensus_hash(&self) -> Hash

Generate a SHA256 hash of the consensus encoding

Can be used to validate all federation members agree on state without revealing the object