Crate fedimint_server::modules::mint

Modules

common

config

db

Structs

BackupRequest

BlindNonce

Nonce but blinded by the user key

Mint

Data structures taking into account different amount tiers Federated mint member mint

MintConsensusItem

A consenus item from one of the federation members contributing partials signatures to blind nonces submitted in it

MintGen

MintGenParams

MintInput

MintOutput

MintOutputBlindSignatures

Result of Federation members confirming MintOutput by contributing partial signatures via MintConsensusItem

MintOutputOutcome

MintOutputSignatureShare

Blind signature share from one Federation peer for a single MintOutput

MintShareErrors

Represents an array of mint indexes that delivered faulty shares

Nonce

Unique ID of a mint note.

Note

An verifiable one time use IOU from the mint.

SignedBackupRequest

VerifiedNotes

Enums

CombineError

MintError

PeerErrorType