Crate fedimint_server::modules::mint

Modules

common

config

db

Structs

BlindNonce

Nonce but blinded by the user key

Mint

Data structures taking into account different amount tiers Federated mint member mint

MintConfigGenParams

MintConfigGenerator

MintInput

MintOutput

MintOutputConfirmation

A consenus item from one of the federation members contributing partials signatures to blind nonces submitted in it

MintOutputOutcome

MintShareErrors

Represents an array of mint indexes that delivered faulty shares

Nonce

Unique ID of a mint note.

Note

An verifiable one time use IOU from the mint.

OutputConfirmationSignatures

Blind signature share from one Federation peer for a single MintOutput

OutputOutcome

Result of Federation members confirming MintOutput by contributing partial signatures via MintOutputConfirmation

VerificationCache

Enums

CombineError

MintError

PeerErrorType