Module fedimint_server::net::connect

Provides an abstract network connection interface and multiple implementations

Modules

mock

Fake network stack used in tests

Structs

PeerCertStore

TlsConfig

TlsTcpConnector

TCP connector with encryption and authentication

Traits

Connector

Allows to connect to peers and to listen for incoming connections

Functions

dns_sanitize

Sanitizes name as valid domain name

parse_host_port

Parses the host and port from a url

Type Aliases

AnyConnector

Owned Connector trait object

ConnectResult

Result of a connection opening future

ConnectionListener

Owned trait object type for incoming connection listeners

SharedAnyConnector

Shared Connector trait object