Module fedimint_server::net::framed

Adapter that implements a message based protocol on top of a stream based one

Structs

BidiFramed

Framed transport codec for streams

BincodeCodec

Framed codec that uses bincode to encode structs with serde support

Traits

FramedTransport

A bidirectional framed transport adapter that can be split into its read and write half

Type Definitions

AnyFramedTransport

Owned FramedTransport trait object

FramedSink

Sink (sending) half of BidiFramed

FramedStream

Stream (receiving) half of BidiFramed

TcpBidiFramed

Special case for tokio TcpStream based BidiFramed instances