Trait fedimint_server::net::framed::FramedTransport

pub trait FramedTransport<T>: Sink<T, Error = Error> + Stream<Item = Result<T, Error>> {
    // Required method
    fn borrow_split(
        &mut self
    ) -> (&mut (dyn Sink<T, Error = Error> + Send + Unpin), &mut (dyn Stream<Item = Result<T, Error>> + Send + Unpin));

    // Provided method
    fn into_dyn(self) -> AnyFramedTransport<T>
       where Self: Sized + Send + Unpin + 'static { ... }
}

A bidirectional framed transport adapter that can be split into its read and write half

Required Methods

fn borrow_split( &mut self ) -> (&mut (dyn Sink<T, Error = Error> + Send + Unpin), &mut (dyn Stream<Item = Result<T, Error>> + Send + Unpin))

Split the framed transport into read and write half

Provided Methods

fn into_dyn(self) -> AnyFramedTransport<T>where Self: Sized + Send + Unpin + 'static,

Transforms concrete FramedTransport object into an owned trait object