Type Alias fedimint_server::net::framed::TcpBidiFramed

pub type TcpBidiFramed<T> = BidiFramed<T, OwnedWriteHalf, OwnedReadHalf>;

Special case for tokio TcpStream based BidiFramed instances

Aliased Type

struct TcpBidiFramed<T> {
    sink: FramedWrite<OwnedWriteHalf, BincodeCodec<T>>,
    stream: FramedRead<OwnedReadHalf, BincodeCodec<T>>,
}

Fields
sink: FramedWrite<OwnedWriteHalf, BincodeCodec<T>>stream: FramedRead<OwnedReadHalf, BincodeCodec<T>>
Implementations

impl<T> TcpBidiFramed<T>where T: Serialize + DeserializeOwned,

pub fn new_from_tcp(stream: TcpStream) -> TcpBidiFramed<T>

Special constructor for tokio TCP connections.

Tokio TcpStream implements an efficient method of splitting the stream into a read and a write half this constructor takes advantage of.

impl<T, WH, RH> BidiFramed<T, WH, RH>where WH: AsyncWrite, RH: AsyncRead, T: Serialize + DeserializeOwned,

pub fn new<S>(stream: S) -> BidiFramed<T, WriteHalf<S>, ReadHalf<S>>where S: AsyncRead + AsyncWrite,

Builds a new BidiFramed codec around a stream stream.

See TcpBidiFramed::new_from_tcp for a more efficient version in case the stream is a tokio TCP stream.

pub fn borrow_parts( &mut self ) -> (&mut FramedSink<WH, T>, &mut FramedStream<RH, T>)

Splits the codec in its sending and receiving parts

This can be useful in cases where potentially simultaneous read and write operations are required. Otherwise a we would need a mutex to guard access.