fedimint_server::net

Module p2p_connector

Provides an abstract network connector interface and multiple implementations

Structs

IrohConnector

TlsConfig

TlsTcpConnector

TCP connector with encryption and authentication

Constants

FEDIMINT_P2P_ALPN ๐

Traits

IP2PConnector

Allows to connect to peers and to listen for incoming connections. Connections are message based and should be authenticated and encrypted for production deployments.

Functions

dns_sanitize

Sanitizes name as valid domain name

gen_cert_and_key

parse_p2p

Parses the host and port from a url

Type Aliases

DynP2PConnector