Module fedimint_server::net::peers

Implements a connection manager for communication with other federation members

The main interface is fedimint_core::net::peers::IPeerConnections and its main implementation is ReconnectPeerConnections, see these for details.

Structs

CommonPeerConnectionState 🔒

ConnectedPeerConnectionState 🔒

DelayCalculator

Calculates delays for reconnecting to peers

DisconnectedPeerConnectionState 🔒

NetworkConfig

Specifies the network configuration for federation-internal communication

PeerConnection 🔒

PeerConnectionStateMachine 🔒

PeerMessage

Internal message type for ReconnectPeerConnections, just public because it appears in the public interface.

PeerStatusChannels

Keeps the references to a PeerStatusChannelSender for each PeerId, which can be used to ask the corresponding PeerConnectionStateMachine for the current PeerConnectionStatus

PeerStatusQuery 🔒

ReconnectPeerConnections

Connection manager that automatically reconnects to peers

Enums

PeerConnectionState 🔒

Constants

PING_INTERVAL 🔒

Every how many seconds to send an empty message to our peer if we sent no messages during that time. This helps with reducing the amount of messages that need to be re-sent in case of very one-sided communication.

Traits

PeerSlice

Type Definitions

PeerConnector

Owned Connector trait object used by ReconnectPeerConnections

PeerStatusChannelReceiver 🔒

PeerStatusChannelSender 🔒